rename index to ids
