+++ packages/frontend/src/components/project/navigation/MobileProjectNavigation.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { ProjectDetailsSection } from '../../../pages/scaling-projects/props/getProjectDetails'
 import { ChevronLeftIcon, ChevronRightIcon } from '../../icons'
-import { MOBILE_PROJECT_NAVIGATION_IDS } from './index'
+import { MOBILE_PROJECT_NAVIGATION_IDS } from './ids'
 
 interface Props {
   sections: ProjectDetailsSection[]
